# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AdminUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      faculty_id:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      last_name:
        type: string
      student_id:
        type: string
      university_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        description: |-
          error message
          example: invalid request format
        type: string
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      faculty_id:
        type: string
      first_name:
        type: string
      gender:
        enum:
        - male
        - female
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      student_id:
        type: string
      university_id:
        type: string
    required:
    - email
    - first_name
    - gender
    - last_name
    - password
    - student_id
    type: object
  dto.ResetPasswordRequest:
    properties:
      password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
...
  dto.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
host: localhost:8080
info:
  contact: {}
  title: Termustat API
  version: "1.0"
paths:
...
  /v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset link if the email exists
      parameters:
      - description: Forgot password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: If the email exists, a reset link will be sent'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to process request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Forgot Password
      tags:
      - auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns a JWT token
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'token: JWT token'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Email not verified
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to login
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and sends a verification email
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Registration successful'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid payload or IDs
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: email or student ID already exists
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to register user
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register
      tags:
      - auth
  /v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using the provided token
      parameters:
      - description: Reset password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Password reset successful'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid payload or token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to reset password
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Reset Password
      tags:
      - auth
  /v1/auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verifies a user's email using the provided token
      parameters:
      - description: Verification payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Email verified successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Failed to verify email
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Verify Email
      tags:
      - auth
...
  /v1/user/me:
    get:
      description: Retrieves information about the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AdminUserResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - auth
swagger: "2.0"